Initialize sentry based on the user's configured telemetry option, and record tracked events as breadcrumbs
The sentry docs say that a list of breadcrumbs with no error is valid; we need to test to make sure this works for just reporting telemetry though.
We should make a list of event name identifiers that the app sends which we should report to sentry at a level other than info; this would make it much easier to distinguish what is actually a problem.