Experimentation with playing previous sound
Variable called previousSounds initalized, though not done properly. Currently, previousSound is set to the current sound every time the code runs, but I only want this to happen when it hasn't been initialized, and I can't figure how to do this without getting errors. Daniel has directed me to another variable that does show the callout history, though this is an array of type POI and not a single variable of sounds. I'm not 100% sure how to utilize this.