Bugfixes, cleanup, moved documentation to docs folder
Fixed bugs with shake device callouts listed in the previous commit. Now, the callout animation properly stops and there is no crash when disabling all callouts.

However, when doing a shake callout on "my location", I noticed that shaking the device will state the callout but also the coordinates, which seems like unnecessary information to the user. I wonder if this is a thing running it on the simulator? Using media controls on my phone to call the previous callout that way doesn't list the coordinates.

I also cleaned up a few more files that I missed the first time around and moved my documentation to the appropriate folder.