Add functionality to repeat last callout upon shaking device + RCOS Documentation (#94)
* Create Soundscape Lab Notebook.pdf

Information about my progress on Soundscape so far.

* Motion detected code + testing code

LaunchHelper.swift contains the code for detecting motion

POICallout.swift contains the sound that I could get to run upon a POI callout.

* More experimentation with playing proper sound

Using a boolean variable to detect motion, a previous sound variable, and reusing existing code, I attempt to play the previous sound when the device is shaken.

* Experimentation with playing previous sound

Variable called previousSounds initalized, though not done properly. Currently, previousSound is set to the current sound every time the code runs, but I only want this to happen when it hasn't been initialized, and I can't figure how to do this without getting errors. Daniel has directed me to another variable that does show the callout history, though this is an array of type POI and not a single variable of sounds. I'm not 100% sure how to utilize this.

* Create Soundscape Lab Notebook.pdf

Forgot to add the updated lab notebook detailing my progress in the previous commit, so I'm adding it to this commit.

* Somehow working code

The code half-accomplishes the goal - shaking the device runs code originally obtained from a function called handleRepeat in HomeViewController+RemoteControl.swift to restart the existing callout. However, when the restarted callout is complete, it will not continue to state other callouts if there are more in the queue.

* Update documentation + change to markdown format

* Minor edit to documentation

Made a line of code appear as a code block.

* Code cleanup, add shake toggle w/ bugs

Shake toggle at its core works but presents two bugs
1. Callout animation does not stop after all callouts have been played
2. Disabling all callouts causes Soundscape to crash stating that "Invalid batch updates detected: the number of sections and/or rows returned by the data source before and after performing the batch updates are inconsistent with the updates."

* Bugfixes, cleanup, moved documentation to docs folder

Fixed bugs with shake device callouts listed in the previous commit. Now, the callout animation properly stops and there is no crash when disabling all callouts.

However, when doing a shake callout on "my location", I noticed that shaking the device will state the callout but also the coordinates, which seems like unnecessary information to the user. I wonder if this is a thing running it on the simulator? Using media controls on my phone to call the previous callout that way doesn't list the coordinates.

I also cleaned up a few more files that I missed the first time around and moved my documentation to the appropriate folder.

* Keep just the changes necessary for this feature

---------

Co-authored-by: Daniel W. Steinbrook <dsteinbrook@gmail.com>